requirements.txt
gradio
python-docx>=1.2.0